feat: use zlib zstd
fix: fix tests
