+++ .github/workflows/tester.yml
@@ -8,10 +8,10 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, windows-latest, macos-latest]
-        node-version: ["18", "20", "22"]
+        node-version: ["22", "24"]
       fail-fast: false
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v5
       - name: Use Node.js ${{ matrix.node-version }}
         uses: actions/setup-node@v4
         with:
@@ -22,42 +22,29 @@ jobs:
           path: node_modules
           key: ${{ runner.os }}-npm-cache
           restore-keys: ${{ runner.os }}-npm-cache
+      - name: Determine unrs-resolver binary version
+        shell: bash
+        run: |
+          case "$RUNNER_OS" in
+            "Linux")
+                echo "PLATFORM=linux-x64-gnu" >> "$GITHUB_ENV" ;;
+            "Windows")
+                echo "PLATFORM=win32-x64-msvc" >> "$GITHUB_ENV" ;;
+            "macOS")
+                echo "PLATFORM=darwin-arm64" >> "$GITHUB_ENV" ;;
+          esac
       - name: Install Dependencies
-        run: npm install
-      - name: Install zstd binary
-        run: npm install "@mongodb-js/zstd"
+        shell: bash
+        run: |
+          npm install
+          npm install "@unrs/resolver-binding-$PLATFORM"
       - name: Test
         run: npm run test
         env:
           CI: true
-  coverage:
-    runs-on: ${{ matrix.os }}
-    strategy:
-      matrix:
-        os: [ubuntu-latest]
-        node-version: ["20.x"]
-    steps:
-      - uses: actions/checkout@v4
-      - name: Use Node.js ${{ matrix.node-version }}
-        uses: actions/setup-node@v4
-        with:
-          node-version: ${{ matrix.node-version }}
-      - name: Install Bun
-        uses: oven-sh/setup-bun@v1
-      - name: Cache NPM dependencies
-        uses: actions/cache@v4
-        with:
-          path: node_modules
-          key: ${{ runner.os }}-npm-cache
-          restore-keys: ${{ runner.os }}-npm-cache
-      - name: Install Dependencies
-        run: bun install
-      - name: Coverage
-        run: npm run test
-        env:
-          CI: true
       - name: Upload coverage report to Codecov
-        uses: codecov/codecov-action@v4
+        if: matrix.os == 'ubuntu-latest' && matrix.node-version == '22'
+        uses: codecov/codecov-action@v5
         with:
           fail_ci_if_error: true
         env:
