+++ lib/js.js
@@ -24,7 +24,7 @@ async function minifyJs(str, data) {
   delete jsOptions.logger
 
   try {
-    const code = await terserMinify(str, jsOptions).code
+    const code = await terserMinify(str, jsOptions).then(x => x.code)
     if (verbose) logFn.call(this, str, code, path, 'js')
     return code
   } catch (err) {
