+++ lib/xml.js
@@ -15,7 +15,7 @@ async function minifyXml() {
   const routeList = route.list()
   const { globOptions, include, verbose } = options
 
-  await Promise.all((match(routeList, include, globOptions)).map(path => {
+  return await Promise.all((match(routeList, include, globOptions)).map(path => {
     return new Promise((/** @type {(value: void) => void} */ resolve, reject) => {
       const assetPath = route.get(path)
       let assetTxt = ''
