+++ lib/zlib.js
@@ -3,6 +3,7 @@ const zlib = require('zlib')
 const { promisify } = require('util')
 const gzip = promisify(zlib.gzip)
 const br = promisify(zlib.brotliCompress)
+const zstd = promisify(zlib.zstdCompress)
 const { match, logFn } = require('./tools')
 
 /**
@@ -81,7 +82,46 @@ function brotliFn() {
   }))
 }
 
+/**
+ * @this {import('@types/hexo')}
+ */
+function zstdFn() {
+  const hexo = this
+  const options = hexo.config.minify.zstd
+
+  const route = hexo.route
+  /** @type {string[]} */
+  const routeList = route.list()
+  const { globOptions, include, verbose, ensureCompressed } = options
+  let level = options.level
+  if (typeof level !== 'number') level = zlib.constants.ZSTD_CLEVEL_DEFAULT
+
+  return Promise.all((match(routeList, include, globOptions)).map(path => {
+    return new Promise((/** @type {(value: void) => void} */ resolve, reject) => {
+      const assetPath = route.get(path)
+      let assetTxt = ''
+      assetPath.on('data', chunk => (assetTxt += chunk))
+      assetPath.on('end', async () => {
+        if (assetTxt.length) {
+          try {
+            const result = await zstd(assetTxt, { params: { [zlib.constants.ZSTD_c_compressionLevel]: level } })
+            const buffer = Buffer.from(assetTxt)
+            if (verbose) logFn.call(this, buffer, result, path, 'zstd')
+            if (!ensureCompressed || buffer.length > result.length) {
+              route.set(path + '.zst', result)
+            }
+          } catch (err) {
+            reject(new Error(`Path: ${path}\n${err}`))
+          }
+        }
+        resolve()
+      })
+    })
+  }))
+}
+
 module.exports = {
   gzipFn,
-  brotliFn
+  brotliFn,
+  zstdFn
 }
