+++ package.json
@@ -13,7 +13,7 @@
   ],
   "scripts": {
     "lint": "standard",
-    "test": "jest"
+    "test": "cross-env NODE_OPTIONS=\"--experimental-vm-modules\" jest"
   },
   "engines": {
     "node": ">= 18.12.0"
@@ -26,29 +26,18 @@
     "url": "git+https://github.com/curbengh/hexo-yam.git"
   },
   "dependencies": {
-    "clean-css": "^5.1.2",
+    "clean-css": "^5.3.3",
     "html-minifier-terser": "^7.2.0",
-    "micromatch": "^4.0.2",
+    "micromatch": "^4.0.8",
     "minify-xml": "^4.5.2",
-    "svgo": "^3.0.0",
-    "terser": "^5.3.0"
+    "svgo": "^4.0.0",
+    "terser": "^5.44.0"
   },
   "devDependencies": {
-    "hexo": "^7.1.0",
-    "jest": "^29.1.2",
-    "standard": "^17.0.0"
-  },
-  "peerDependencies": {
-    "@mongodb-js/zstd": "^2.0.0",
-    "@bokuweb/zstd-wasm": "^0.0.22"
-  },
-  "peerDependenciesMeta": {
-    "@mongodb-js/zstd": {
-      "optional": true
-    },
-    "@bokuweb/zstd-wasm": {
-      "optional": true
-    }
+    "cross-env": "^10.1.0",
+    "hexo": "^8.1.1",
+    "jest": "^30.2.0",
+    "standard": "^17.1.2"
   },
   "keywords": [
     "minify",
